+++ app/Policies/TirePolicy.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Policies;
+
+use App\Models\Tire;
+use App\Policies\Concerns\HasGroupScopedAccess;
+
+class TirePolicy
+{
+    use HasGroupScopedAccess;
+}